abstracts out a generic approach to identifying array callback method calls
This commit pulls out the components of the previous version which were not specific to `forEach` and instead makes it possible to refer to all method applications instead of just `forEach` applications. To refer to specific methods, we instead just have an accessor that returns the method name, which can be constrained in the query's `where` clause.