cleans up function library and enable indirect reference to functions
This commit improves the representation of method applications by exposing the
arguments of the method, which makes it possible to refer to them directly instead of using the subclasses that pulled out the callbacks. This also adds a new abstraction on top of functions to allow indirect reference to functions via variables.